gem: Use CcsdsTransferFrames namespace
Use a proper namespace for the gem, which seems like a more common way
to structure gems in general, this also means we are less likely to
conflict with some existing custom protocol.

The con is that additional and more explicit configuration is needed in
order to expose the protocol for use in a target (noted in README.md).