ccsds-tf: Make FHP take precedence over pkt length
According to the CCSDS standard:

4.3.2.4 If the calculated location of the beginning of the first Packet
is not consistent with the location indicated by the First Header
Pointer, then the Packet Extraction Function shall assume that the First
Header Pointer is correct, and shall continue the extraction based on
that assumption.

Change the packet continuation handling in order to:

* Give the first header pointer precedence when the length of the
  continued packet indicated overlap. In this case the continued packet
  will be cut short.

* Satisfy both the continued packet length and the first header pointer
  if they indicate a gap between the continued packet end and the start
  of the next packet, by discarding the data in the gap.

The removal of the data before the first header pointer is now fully
handled by handle_packet_continuation().

The check if further processing should be done by store_packets() is now
moved into process_frame(), and handle_packet_continuation() can no
longer return a :STOP symbol for this reason.

Tests are added for some first header pointer and packet length
inconsistency scenarios.