ccsds-tf: Avoid temp. copy for BinaryAccessor
Rework calls to BinaryAccessor for getting space packet apid and length
to avoid an intermediate copy of the source string.

Move these actions into inline calls instead, since the lack of copying
makes it less natural to encapsulate the call in a generic function.

Remove assert-like exceptions for checking that the size is sufficient
for the BinaryAccessor call, since the BinaryAccessor already provides
these exceptions.

Rough measurements indicates that this gives a slight performance
increase.